app/gateway/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
soundfile==0.13.1
numpy
tritonclient
packaging